standalone: modified prototypes of xil_In32 and xil_Out32
Following changes have been done to remove the warnings,
1.prototype of xil_In32 and xil_Out32 is modified.Changed type
  of argument from "INTPTR" to "UINTPTR".
2.Argument of Xil_ExceptionEnableMask is typecasted to
  u32 while function call.

+++ lib/bsp/standalone/src/cortexa9/xil_exception.h
@@ -127,7 +127,7 @@ typedef void (*Xil_InterruptHandler)(void *data);
 *
 ******************************************************************************/
 #define Xil_ExceptionEnable() \
-		Xil_ExceptionEnableMask(XIL_EXCEPTION_IRQ)
+		Xil_ExceptionEnableMask((u32)XIL_EXCEPTION_IRQ)
 
 /****************************************************************************/
 /**
